+++ src/graphnet/data/sqlite/__init__.py
@@ -1,4 +1,4 @@
 from .sqlite_dataconverter import SQLiteDataConverter
 from .sqlite_dataset import SQLiteDataset
 from .sqlite_dataset_perturbed import SQLiteDatasetPerturbed
-from .sqlite import run_sql_code, save_to_sql
+from .sqlite_utilities import run_sql_code, save_to_sql
